feat: standalone video link
