+++ src/components/Common/StandaloneVideoLink.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import BaseVideoLink from './BaseVideoLink'
+
+const StandaloneVideoLink = props => (
+  <BaseVideoLink mode="standalone" {...props} />
+)
+
+export default StandaloneVideoLink
